Rename things
kcmp -> kmp
Diff -> SafeDiff
